+++ test/integration/068_partial_parsing_tests/extra-files/gsm_override.sql
@@ -1,0 +1,6 @@
+- custom macro
+{% macro generate_schema_name(schema_name, node) %}
+
+    {{ schema_name }}_{{ target.schema }}
+
+{% endmacro %}